+++ Gruntfile.js
@@ -163,7 +163,6 @@ module.exports = function (grunt) {
                 src: [
                     '<%= conf.web %>/js/d3.js',
                     '<%= conf.web %>/js/crossfilter.js',
-                    '<%= conf.web %>/js/colorbrewer.js',
                     '<%= conf.pkg.name %>.js'
                 ]
             }
@@ -182,7 +181,6 @@ module.exports = function (grunt) {
                     // JS code dependencies
                     '<%= conf.web %>/js/d3.js',
                     '<%= conf.web %>/js/crossfilter.js',
-                    '<%= conf.web %>/js/colorbrewer.js',
                     // Code to be tested
                     '<%= conf.pkg.name %>.js',
                     // Jasmine spec files
@@ -227,7 +225,25 @@ module.exports = function (grunt) {
                 }
             },
             ci: {
-                browsers: ['Chrome', 'Firefox'],
+                browsers: ['ChromeNoSandboxHeadless', 'FirefoxHeadless'],
+                customLaunchers: {
+                    // See https://github.com/karma-runner/karma/issues/2603
+                    ChromeNoSandboxHeadless: {
+                        base: 'Chrome',
+                        flags: [
+                            '--no-sandbox',
+                            // See https://chromium.googlesource.com/chromium/src/+/lkgr/headless/README.md
+                            '--headless',
+                            '--disable-gpu',
+                            // Without a remote debugging port, Google Chrome exits immediately.
+                            ' --remote-debugging-port=9222'
+                        ]
+                    },
+                    FirefoxHeadless: {
+                        base: 'Firefox',
+                        flags: ['-headless']
+                    }
+                },
                 concurrency: 1,
                 reporters: ['dots', 'summary']
             },